Break up title and tabset helper, start adjusting tests. Many new methods
