+++ app/views/controllers/articles/index.html.erb
@@ -1,6 +1,6 @@
 <%
 @container = :wide
-add_index_title(@query, no_hits: :ARTICLES.t)
+add_index_title(@query)
 add_tab_set(articles_index_tabs(user: @user))
 add_sorter(@query, articles_index_sorts)
