+++ app/views/controllers/collection_numbers/index.html.erb
@@ -1,7 +1,7 @@
 <%
 @container = :wide
 # the no_hits_title.
-add_index_title(@query, no_hits: :list_collection_numbers_title.t)
+add_index_title(@query)
 add_tab_set(collection_numbers_index_tabs(obs: @observation))
 add_sorter(@query, collection_numbers_index_sorts)
