+++ app/views/controllers/herbaria/index.html.erb
@@ -1,6 +1,6 @@
 <%
 @container = :wide
-add_index_title(@query, no_hits: :list_herbaria_title.t)
+add_index_title(@query)
 add_tab_set(herbaria_index_tabs(query: @query))
 add_sorter(@query, herbaria_index_sorts(query: @query))
