+++ app/views/controllers/locations/index.html.erb
@@ -1,6 +1,6 @@
 <%
 @container = :full
-add_index_title(@query, no_hits: "")
+add_index_title(@query)
 add_tab_set(locations_index_tabs(query: @query))
 all_links = if (params[:id].present? ||
                 params[:by].present? ||
