+++ app/views/controllers/projects/index.html.erb
@@ -1,5 +1,5 @@
 <%
-add_index_title(@query, no_hits: :list_projects_title.t)
+add_index_title(@query)
 add_tab_set(projects_index_tabs)
 add_sorter(@query, projects_index_sorts)
