+++ app/views/controllers/sequences/index.html.erb
@@ -1,5 +1,5 @@
 <%
-add_index_title(@query, no_hits: :SEQUENCES.t)
+add_index_title(@query)
 add_sorter(@query, sequences_index_sorts)
 flash_error(@error) if @error && @objects.empty?
 %>
