+++ app/views/controllers/species_lists/index.html.erb
@@ -1,26 +1,26 @@
 <%
-add_index_title(@query, no_hits: "")
+add_index_title(@query)
 add_sorter(@query, species_lists_index_sorts(query: @query))
 flash_error(@error) if @error && @objects.empty?
 %>
 
 <%= paginate_block(@pages) do %>
-<% if @objects.any? %>
-<div class="list-group">
-  <% @objects.each do |species_list| %>
-  <div class="list-group-item">
-    <div class="text-larger">
-      <%=
-              when_span = tag.span(species_list.when.to_s, class: "list_when")
-              what_span = tag.span(species_list.unique_text_name.t,
-                                   class: "list_what")
-              link_to(when_span + ": " + what_span, species_list.show_link_args)
-            %>
-    </div><!-- .text-larger -->
-    <span><%= species_list.place_name.t rescue :UNKNOWN.t %></span> |
-    <span><%= user_link(species_list.user) %></span>
-  </div>
+  <% if @objects.any? %>
+    <div class="list-group">
+      <% @objects.each do |species_list| %>
+      <div class="list-group-item">
+        <div class="text-larger">
+          <%=
+            when_span = tag.span(species_list.when.to_s, class: "list_when")
+            what_span = tag.span(species_list.unique_text_name.t,
+                                class: "list_what")
+            link_to(when_span + ": " + what_span, species_list.show_link_args)
+          %>
+        </div><!-- .text-larger -->
+        <span><%= species_list.place_name.t rescue :UNKNOWN.t %></span> |
+        <span><%= user_link(species_list.user) %></span>
+      </div>
+      <% end %>
+    </div>
   <% end %>
-</div>
-<% end %>
 <% end %>
