+++ app/views/controllers/users/index.html.erb
@@ -1,7 +1,6 @@
 <%
 @container = :full
-no_hits = in_admin_mode? ? :users_by_name_title.t : :title_for_user_search.t
-add_index_title(@query, no_hits: no_hits)
+add_index_title(@query)
 add_sorter(@query, users_index_sorts(in_admin_mode?))
 
 flash_error(@error) if @error && @objects.empty?
