+++ test/classes/query/comments_test.rb
@@ -26,9 +26,9 @@ class Query::CommentsTest < UnitTestCase
   def test_comment_for_target
     obs = observations(:minimal_unknown_obs)
     expects = Comment.index_order.where(target_id: obs.id).distinct
-    assert_query(expects, :Comment, target: { id: obs, type: :Observation })
+    assert_query(expects, :Comment, target: { type: :Observation, id: obs })
     expects = Comment.index_order.target(obs).distinct
-    assert_query(expects, :Comment, target: { id: obs, type: :Observation })
+    assert_query(expects, :Comment, target: { type: :Observation, id: obs })
   end
 
   def test_comment_for_user
