+++ test/controllers/herbarium_records_controller_test.rb
@@ -23,7 +23,7 @@ class HerbariumRecordsControllerTest < FunctionalTestCase
     get(:index)
 
     assert_response(:success)
-    assert_displayed_title("Fungarium Records by Name")
+    assert_displayed_title("Fungarium Records")
     # In results, expect 1 row per herbarium_record
     assert_select("#results tr", HerbariumRecord.count,
                   "Wrong number of Herbarium Records")
@@ -34,8 +34,7 @@ class HerbariumRecordsControllerTest < FunctionalTestCase
 
     login
     get(:index, params: { by: by })
-
-    assert_displayed_title("Fungarium Records by Fungarium")
+    assert_sorted_by(by)
   end
 
   def test_index_by_initial_determination
@@ -43,8 +42,7 @@ class HerbariumRecordsControllerTest < FunctionalTestCase
 
     login
     get(:index, params: { by: by })
-
-    assert_displayed_title("Fungarium Records by Initial Determination")
+    assert_sorted_by(by)
   end
 
   def test_index_by_accession_number
@@ -52,8 +50,7 @@ class HerbariumRecordsControllerTest < FunctionalTestCase
 
     login
     get(:index, params: { by: by })
-
-    assert_displayed_title("Fungarium Records by Accession Number")
+    assert_sorted_by(by)
   end
 
   def test_index_pattern_with_multiple_matching_records
@@ -64,7 +61,8 @@ class HerbariumRecordsControllerTest < FunctionalTestCase
     get(:index, params: { pattern: pattern })
 
     assert_response(:success)
-    assert_displayed_title("Fungarium Records Matching ‘#{pattern}’")
+    assert_displayed_title("Fungarium Records")
+    assert_displayed_filters("pattern: #{pattern}")
     # In results, expect 1 row per herbarium_record
     assert_select("#results tr", 2)
   end
@@ -85,10 +83,9 @@ class HerbariumRecordsControllerTest < FunctionalTestCase
     login
     get(:index, params: { herbarium: herbarium.id })
 
-    assert_displayed_title(
-      :query_title_in_herbarium.l(types: :HERBARIUM_RECORDS.l,
-                                  herbarium: herbarium.name)
-    )
+    assert_displayed_title("Fungarium Records")
+    assert_displayed_filters("fungaria: #{herbarium.name}")
+
     # In results, expect 1 row per herbarium_record
     assert_select("#results tr",
                   HerbariumRecord.where(herbarium: herbarium).count)
@@ -110,10 +107,8 @@ class HerbariumRecordsControllerTest < FunctionalTestCase
     login
     get(:index, params: { observation: obs.id })
 
-    assert_displayed_title(
-      :query_title_for_observation.l(types: :HERBARIUM_RECORDS.l,
-                                     observation: obs.unique_text_name)
-    )
+    assert_displayed_title("Fungarium Records")
+    assert_displayed_filters("observations: #{obs.id}")
     #  "Fungarium Records attached to ‘#{obs.unique_text_name}’")
     assert_select("#results tr", obs.herbarium_records.size)
   end
